add a new file for a class
